+++ addon/index.js
@@ -1,9 +1,4 @@
-export {
-  runTask,
-  scheduleTask,
-  throttleTask,
-  cancelTask,
-  cancelDebounce,
-} from './utils/tasks';
+export { runTask, scheduleTask, throttleTask, cancelTask } from './utils/tasks';
 export { pollTask, pollTaskFor, setShouldPoll, cancelPoll } from './poll-task';
+export { debounceTask, cancelDebounce } from './debounce-task';
 export { runDisposables } from './utils/disposable';
